make visibility work somewhat
